+++ gmx_reader.h
@@ -32,7 +32,7 @@ class gmx_reader
         void find_frame(int frame);
         void read_frame(int frame);
         void index_frames();
-        int  get_frame_number(float time);
+        int  get_frame_number(double time);
         bool checktime(float time);
         // >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
